requirements.txt
streamlit
langchain==0.3.1
langchain-community==0.3.1
langchain-core>=0.3.10
langchain-google-genai
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf2
python-dotenv
sentence-transformers
google-generativeai
